Fix public events not being pulled
This is a quick fix. I assumed that the events are the same across any
year, but this is not the case for various countries. So instead it should
remove the events that exist in both years and then append the rest from
the next year. But that's for later.